+++ src/core/Cone.cpp
@@ -1,4 +1,15 @@
+#include "pch.hpp"
 #include "Cone.hpp"
+#include "Config.hpp"
+
+#include <algorithm>
+#include <cmath>
+#include <string>
+#include <utility>
+#include <vector>
+
+#include <glm/fwd.hpp>
+#include <glm/detail/func_geometric.inl>
 
 void Cone::_generate(const unsigned int segments, const float height, const float radius, const ValuesRange range, const bool useFlatShading)
 {
